tidy up

+++ faba/src/hypothesis_tests.rs
@@ -26,7 +26,7 @@ where
             self.observed.0.to_u64(),
         ) {
             let null_pr = (exp1 as f64) / ((exp0 + exp1) as f64).max(1.0);
-            let nobs = (obs1 + obs0) as u64;
+            let nobs = obs1 + obs0;
             let distrib = Binomial::new(null_pr, nobs)?;
             Ok(1.0 - distrib.cdf(obs1.saturating_sub(1)))
         } else {
@@ -41,7 +41,7 @@ where
             self.observed.0.to_u64(),
         ) {
             let null_pr = (exp1 as f64) / ((exp0 + exp1) as f64).max(1.0);
-            let nobs = (obs1 + obs0) as u64;
+            let nobs = obs1 + obs0;
             let distrib = Binomial::new(null_pr, nobs)?;
             Ok(distrib.cdf(obs1))
         } else {
